feat: separate each component in their own file
