+++ my-app/src/index.css
@@ -6,7 +6,7 @@ body {
   ol, ul {
     padding-left: 30px;
   }
-  
+
   .board-row:after {
     clear: both;
     content: "";
