+++ my-app/src/square-component/squareComponent.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function Square(props) {
+  return (
+    <button className="square" onClick={props.onClick}>
+      {props.value}
+    </button>
+  );
+}
